django_app/requirements.txt
Django==5.1
django-cors-headers==4.2.0
django-filter==24.3
django-rest-framework==0.1.0
django-rest-framework-keycloak==0.2.0
djangorestframework==3.14.0
drf-oidc-auth==3.0.0
ecdsa==0.18.0
exceptiongroup==1.2.2
frozenlist==1.4.0
future==0.18.3
hexbytes==1.2.1
idna==3.4
iniconfig==2.0.0
josepy==1.13.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
mozilla-django-oidc==3.0.0
multidict==6.0.4
numpy==2.0.0
packaging==23.1
pandas==2.2.2
parsimonious==0.10.0
pluggy==1.5.0
protobuf==4.24.2
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.8.2
pydantic_core==2.20.1
pyjwkest==1.4.2
pyOpenSSL==23.2.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-keycloak==3.3.0
pytz==2023.3
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rlp==4.0.1
rpds-py==0.10.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toolz==0.12.0
types-requests==2.32.0.20240712
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.0.4
web3==7.0.0
websockets==11.0.3
yarl==1.9.2
